Restructure the project to make it a package
