+++ README.md
@@ -1,2 +1,31 @@
 # ArtworkRetrievalDataGen
+
 Artwork image retrieval synthetic data generator
+
+## Installation
+
+You will need to have `blender` 3.5 installed, then clone the repo.
+
+```bash
+git clone https://github.com/Oxid15/ArtworkRetrievalDataGen.git
+```
+
+## Usage
+
+1. Create the folder with source images and the one for results
+
+    ```bash
+    mkdir src
+    mkdir dst
+    ```
+
+2. Adjust parameters in the file `settings.py`
+3. Run the rendering
+
+    ```bash
+    <path-to-your-blender>/blender -b -P <path-to-this-repo>/ArtworkRetrievalDataGen/generate.py
+    ```
+
+4. Review the results in the destination folder
+
+    ![Alt text](./examples/00001.png)
